setup: Move into proper package and add additional requires fields
Removes executable bit from `__init__.py` (was `pytop.py`)
Adds `__main__.py` entrypoint
