+++ pytop/__main__.py
@@ -1,0 +1,4 @@
+from pytop import main
+
+if __name__ == '__main__':
+    main()
