+++ pytop/sysinfo.py
