+++ setup.py
@@ -4,6 +4,15 @@ from setuptools import setup, find_packages
 setup(
     name="pytop",
     version='1.0.0',
+    packages=['pytop'],
+    install_requires=[
+        'urwid>=2.1.0',
+    ],
+    extra_requires={
+        "test": [
+            'pytest>=5.3.1',
+        ],
+    },
     description='Htop copycat implemented in Python.',
     author='Andrii Oshtuk',
     url='https://github.com/AndriiOshtuk/pytop',
